pyproject.toml
[tool.poetry]
name = "pb-discord"
version = "0.1.0"
description = ""
authors = ["Project Borealis"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["speedup"], version = "^2.4.0"}
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
coloredlogs = "^15.0.1"
aiohttp = "^3.9.5"
aiodns = "^3.2.0"
colorama = { version = "^0.4.6", markers = "sys_platform == 'win32'" }
rapidfuzz = "^3.9.4"
arrow = "^1.3.0"
cachetools = "^5.4.0"
orjson = "^3.10.6"
tinydb = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
